requirements.txt
Pillow
pytest
flake8
black
